Trait staking_rewards_rpc::StakingRewardsApi

pub trait StakingRewardsApi<BlockHash, Balance>: Sized + Send + Sync + 'static {
    fn yearly_emission(
        &self, 
        total_staked: Balance, 
        total_issuance: Balance, 
        at: Option<BlockHash>
    ) -> Result<Balance>;

    fn max_yearly_emission(&self, at: Option<BlockHash>) -> Result<Balance>;

    fn to_delegate<M: Metadata>(self) -> IoDelegate<Self, M>
    where
        BlockHash: Send + Sync + 'static + DeserializeOwned,
        Balance: Send + Sync + 'static + Serialize + DeserializeOwned,
    { ... }
}

Required methods

fn yearly_emission(
&self,
total_staked: Balance,
total_issuance: Balance,
at: Option<BlockHash>
) -> Result<Balance>

Emission reward 1 year from now given the currently staked funds and issuance. Depends on the reward curve, decay percentage and remaining emission supply.

fn max_yearly_emission(&self, at: Option<BlockHash>) -> Result<Balance>

Maximum emission reward for 1 year from now. Depends on decay percentage and remaining emission supply.

Provided methods

fn to_delegate<M: Metadata>(self) -> IoDelegate<Self, M> where
BlockHash: Send + Sync + 'static + DeserializeOwned,
Balance: Send + Sync + 'static + Serialize + DeserializeOwned,

Create an IoDelegate, wiring rpc calls to the trait methods.